Tab related changes: - Custom tabs now use NSTabViewItem's drawLabel and labelSize methods - The message window has been modified to use a custom subclass of NSTabViewItem for it's message tabs.
As a result of these changes, AICustomTabsView draws it's tab labels the same way as NSTabView.
This means that any custom tab label content (colors, icons) will display the same way whether
an AICustomTabsView is handling the tabs, or just a stock NSTabView.  It also greatly simplifies
the task of creating custom tab label content.