**Modifications to how account status is treated**
Instead of requiring a separate protocol, methods, code, and notification for each aspect of an
account's status (away, idle, profile, online/offline), I've combined them all together using a
status dictionary (almost identical to how contact's status is treated).

AIAccountController now responds to:
- (id)statusObjectForKey:(NSString *)key account:(AIAccount *)inAccount;
- (void)setStatusObject:(id)inValue forKey:(NSString *)key account:(AIAccount *)inAccount;

These method provide access to the status dicts, and should always be used (even by the accounts
themselves).

Notes on the 'setStatus' method:
- Passing a nil account will set the value for all accounts that lack a specific value.
- Passing a nil value with a valid account will remove an account-specific value.

----
I've modified the existing account code to recognize the following keys:
"Online"		YES/NO
 "Status"		Offline/Connecting/Online/Disconnecting (read only)
"IdleTime"		# Seconds
"IdleManuallySet"	YES/NO

When a value is changed, an Account_StatusChanged notification is broadcast.  Changing an
account's status is as simple as setting the various keys.

These changes greatly reduce the work of future account status additions, and should help clean
up the account implementation quite a bit.

Several account protocols have been removed and are no longer needed.

----
I've added an 'availableForSendingContentType:toHandle:' method that should be used in place of
checking an account's online status (as a means of determining if an account is available for
sending content).

The existing 'accountForSendingContentToHandle' method has been modified to require a content
type, allowing for a unique account choice depending on the type of content being sent.

---
I've left in some status change logging which I feel is necessary for now, but will be removed
in a bit.