Removed the syncmail test warnings.
Incase anyone is wondering, syncmail is the little script that sends out emails whenever a cvs
commit occurs.  It wasn't including a diff in the emails (and was generating errors) when one of
our folder names contained a space (which is almost every folder we have :P).  It appears to be
fixed now though.