I've changed how the build targets in the Adium project work, resulting in greatly increased compile time.
When working on a plugin, select it as the active target.  With this new setup, the plugin will
be built, and automatically copied into the built Adium.app.

(With the previous setup, plugins were not copied until the main Adium target was built,
requiring us to build and re-link it every compile)
----
AIFlexibleTableView (and away messages):
- Flexible cells are now inverted when selected
- Deleting the selected row no longer causes problems
- Deleting the edited row no longer causes problems
- Cells correctly expand & shrink vertically while being edited
- Seperating cell into text & image
- Replaced the [:)] with a (temp) away message icon

Other changes:
- Put sounds back into /contents/resources/
- Plugins are now iconed & bundled (you may need to log out or place Adium 2.0 in your apps
folder for this to take effect)