- Sounds now play much quicker. - Adium no longer blocks while sound hardware powers up.
These changes should greatly reduce the lag experienced when sounds play, especially so for
laptop users. ([ 722639 ] Adium blocks while playing sounds)

When sound hardware is powered down and a sound is played, there is still a noticable delay
before the sound starts.  There is no way around this.  The difference is that now Adium remains
responsive during this delay.

- I've added the ability to choose a custom volume level.
- Also added the ability to mute sounds.

Setting a custom volume switches Adium back to the old method of playing sounds, with slower
performance and hardware powerup delays.  There is no way around this.