Forgot to fix a line I'd commented out for testing.
Currently, 'editable' and 'enabled' mean the same thing to the contact list editor.  Technically, the
oscar account list should be enabled now but not 'editable' - however, setting editable to NO also
disables the account in the editor.  I guess for a fully implementeed service, they should be the same
thing, though, so it's probably not worth worrying about.

----------------------------------------------------------------------