- Fixed a crash on quit caused by the contact list window.
(We're getting this crash because of a flaw in how interface containers are setup.  Things should be
changed so that the interface is the only one to initiate the closing of a container, instead of
containers closing themselves and notifying the interface.  It'd be nice to change this in the future,
but until then my temporary fix will keep us from crashing here)