Added a build number to the about box. click on the build date to toggle between the date and the number, ala About This Mac...
Anyway, buildnum is under the MIT license (identical to the BSD license
w/o the advertising clause). It needs to be a separate program to get
around the problem of "incremental builds." The code is provided as
reference, you don't need to compile it, and it shouldn't be compiling.

This took me a couple days to come up with, I hope you guys like it :D