Ahh, wonderful libgaim. More improvements to how things are organized. Everything is peachy. I really wish I could silence some of those warnings that pop up though.
For the technical: I've moved libgaim UiOp functions 100% in to the
service plugin.everything seems to be working well with that.
CBGaimAccount (which is either going to be renamed to CBGaimOscarAccount or
subclassed, I don't know which) is now totally unaware of the UiOp
functions. It does deal with the libgaim classes (which is a Good Thing,
because CBGaimServicePlugin should only be forwarding events to the
accounts). It's quite elegant, I think. Let me know if you have any
comments. I'd like you guys to keep your hands out of this for a little
while longer...sorry! :'( I wanna get the buddy list totally working, so
you have something to write off of.

Alright, that's enough of this massive commit message. Take care, and
I'll see you on IRC :)