Added the files neccesary to build a public binary.
-buildDMG.pl will take care of packaging the files
-cvs2cl Will run twice and log the changes

****WARNING****
This script will checkout adium each and everytime.  IT WILL DELETE IT before it checks it out again (to make sure everything is good).  So don't develop with it please I would hate to hear how someone spent 5 hours and then wanted to build a public binary to show it off. ;-)