- Added away message support to liboscar/libgaim.
+ Added commented out code which should do the Signon Date, but doesn't.  There's a general version in CBGaimAccount and an aim-specific
version in CBGaimAIMAccount... neither works, though either should.

+ Added liboscar.a to the project (xcode only - someone needs to add it to the gaim target in project builder).  It is used for reference by
xcode to resolve a bunch of warnings which come from using oscar-specific functions; it is not in the copy phase because it is not needed in
the built product, as liboscar.so contains the exact same information, just in a format we can't access from xcode.