Updated the build scripts.  All three files should be in a folder which contains the checked out 'adium' folder, like so:
drwxr-xr-x  17 evands  staff    578  3 Nov 00:19 adium
-rwxr-xr-x   1 evands  staff   1742  3 Nov 00:18 build.sh
-rwxr-xr-x   1 evands  staff  12356  1 Nov 04:52 buildDMG.pl
-rwxr-xr-x   1 evands  staff  77975 31 Oct 03:20 cvs2cl.pl

I have it set to run a 4 am every night on my local machine; these scripts exist in the repository just for reference.  If you want to use
them, be sure to take out the lines near the end which actually upload the files to sourceforge (unless that's what you want to be doing, of course).