Rough draft of code to adapt libgaim to Cocoa event loop. Not used yet.
This requires glue code on libgaim side, too. That patch is also in the
rough draft stages. Once it's finished, I'll have to beg someone else to
compile it because a compile on my Panther machine won't work for the poor
Jaguar people.

This is very self-contained, so in theory other Cocoa- and libgaim-based
projects could use it.