- AccountController now has methods for setting an account's userIcon and for setting the default user icon for accounts which don't specify one.  This can't use the normal setProperty: forKey: stuff because we can't write NSImages to the .plist file when the account data saves, and caching the NSImage is -well- worth it. - AIAccount was also expanded to handle userIcons. - GaimAccount handling of its own userIcon fixed up and refined. - Address Book Integration grabs the "me" card if available and set its image, if available, as the default user icon.
User icons are cached as jpg images in the ~/Library/Caches/Adium folder.  Buddy images might do well to be cached, as well, so we don't have to wait for the server to supply them... I may write a plugin to use the same techniques.
User icons must be cached because libgaim certainly doesn't want an NSImage... it's wants a local file name, and we want to be able to let plugins supply an arbitrary image without having to worry about where its coming from or how long it'll be there.