Revamped the implementation of object specific preferences. ** All current contact-specific preferences will be lost, sorry.
Technical details:
Object specific preferences are now loaded, saved, and controlled by the actual objects.
Retrieving an object preference is a simple preferenceForKey:group: .  They are now saved in
a 'ByObject' folder, with a separate file for each contact.

If the object does not have a value for the preference, it will return the value of it's
containing group.  This will traverse back to the root of the contact list, where the last
object with no containing group will return Adium's global value for the preference.  This
traversal is fairly intelligent and shouldn't be that much slower than a regular preference
lookup.

This means that we now have a true implementation of object specific preferences, and
preferences can be applied to groups (to affect all contained contacts) in addition to
contact objects.

The two early adopter plugins (Announcer and contact alerts) had their own implementation of
object preferences, so they were affected by this change.  I've done my best to correct both
plugins, but I believe the contact alert preferences are having issues with group
preferences.  I'm going to take a look at this next.

Contact list alias editing was also broken by this change, but this will be corrected by the
planned changes to the contact list editor that are coming later on.