Made some major changes to how accounts work behind the scenes.
Technical details:
AIAccount is now a subclass of AIListObject.  This change allows it to re-use a lot of he
list object code instead of it's own.  Account properties are now stored as account specific
preferences.  When appropriate, account code adds status keys to the account object as if it
were a contact.  These changes are quite beneficial, but their benefits will be seen indirectly
through upcoming changes.

Since accounts are now a subclass of list object, the contact list can how handle displaying
them alongside buddies... this may prove interesting later on.

Account preferences are now stored in separate files, one per account.  In addition, deleting
an account no longer destroys it's custom preferences... which makes it quite difficult to
destroy a user's account information (good thing).

There is temporary code to import existing accounts into the new layout.  Everything should
continue to work smoothly, but be on the lookout for new bugs, especially with areas of Adium
dealing with accounts.