Let's try this fix instead for the status issues.  It appears that the main problem was accounts being told to connect at the same time - so this would manifest itself when using connect all or multiple auto-connect's on launch.  Accounts now correctly ignore duplicate password returns (Instead of attempting to connect twice like before).
I've also changed things so that status changes meant for another account
are ignore by everyone else in AIAccount as soon as possible.

Removed the delay on setting our initial status with libgaim as well.  We
don't appear to need this anymore (I believe it was because of the above
issues that it was required in the first place)