In case it cuts off: http://outpost.penguinmilitia.net/~evands/12-28_commit.txt
- Fixed a minor bug preventing the crash reporter from properly loading exception logs

- Added local display name and profile setting to AIM via libgaim.

- Fixed a crash on the account preference view which could only occur if no accounts were present.

- Added a sort function to the services list in the account preference view to sort services alphabetically.

- Couldn't decide which of the protocols to implement next, so did all of them we have access to at the moment *grin*:
MSN
Yahoo
Gadu-Gadu
Napster
Jabber

PLEASE don't flood the forums with "When will MSN chat support be available?" and other protocol-specific questions!  It'll happen when it can :)

The binary is going to be a bit bigger now - MSN requires a whole suite of SSL libraries in order to connect.  It's worth it though!  W00t - Adium r0x!

-----
Technical information:
CBGaimServicePlugin is no longer technically a service plugin in Adium's terms... however, it does provide the service of bridging us to Gaim, so the name is still appropriate.
After initializing Gaim, it calls the GaimServices to do the actual Adium work.
The 6 gaim protocols ("prpls" as they call them, protocol plugins) are implemented through subclasses of GaimService, which does most of the work for them.
Each service has its own subclass of CBGaimAccount and ESGaimAccountViewController, and can there do custom handling of whatever it needs.

ESGaimAccountViewController, itself an AIAccountViewController sublcass, provides the capability for the gaim services to specify parameters for account configuration and provide configuration tabs for protocol-specific things; these tabs are added to a single, unified gaim service account pane.
An example of this is seen in the current implementation of AIGaimAIMAccountViewController.

+ Generalized some code in AIAccountViewController away from being TOC2-specific.
+ Touched the TOC2 nib while fixing a spelling error in a variable.  Aren't car rides fun?
+ Added a bunch of debug code to the address book plugin in an attempt to hunt down an elusive bug - don't worry, your console will only be flooded for a day or two.

And I think I'm done for the day. :-)