A lot of contact list opacity cleanup, this should fix all the shadowing issues. We no longer refresh shadows or set the window as transparent unnecessarily.
Unfortunately:
Shadowing is going to be a major CPU hit and there's not much we can do about it.  Panther
broke the trick we were using previously to prevent shadowing of our window's content, which
means that every single contact list update now requires a recalculation of the window
shadow - even with transparency set to where the shadows are barely visible.  Even terminal
is having display issues with transparency in Panther.