Fixed syncmail as per sourceforge.net:
"( 2004-01-23 06:31:27 - Project CVS Service ) ?? Due to the recent OS upgrade on the SourceForge.net developer CVS server, it will be necessary for any per-project syncmail script installations to make a one-line change to their syncmail script. The central copy of the syncmail script which is used by a number of projects has already been modified in this manner. To function properly, the MAILHOST line should be hardcoded to MAILHOST = 'localhost'. Additional information on syncmail may be found in Enhancing Project CVS Services Using Scripts; the SourceForge.net team maintains a central copy of the syncmail script which your project may use."