Is there a legit reason that we allow loading of non-standard plugins?
I've cleaned up the loading code and disabled the external plugins folder
for now.  The issue is that people download non-standard plugins and then
forget they are installed (or neglect to associate them with crashes and
other issues).  It's frustrating to see crashes and issues caused by
outdated plugins being used from the auxilary folder.