Added a 'Script' menu for the iTunes scripts.  The main goal here is to provide an easy way to see the available scripts ... but it's also possible to select them from the menu to insert the text keyword.
I've also changed how the scripts are loaded.  They're now loaded dynamically from
a 'scripts' directory within the plugin bundle.  The menu should handle directories
and subdirectories, correctly labeling them in the menu and indenting the script
names.

Along with this, I've removed the iTunes integration preferences.  There is no
reason to disable this feature, and we no longer need the keyword list in there. :)