Added ESWebFrameViewAdditions to the WebKit plugin.  At present, this just provides access to turning off horizontal scrolling in the web kit frame - the header lists many other private methods.
This is being added to the webkit plugin rather than to AIUtilities for two reasons.  First, I did not want to link AIUtilities against WebKit.framework unnecessarily, as this would needlessly limit the portability of AIUtilities.  Second, accessing the private webkitframe class requires compiling with -undefined suppress since we know it's there but the compiler doesn't.
-undefined suppress requires -flat_namespace, and I'd rather keep AIUtilities with the faster-loading two-level namespace.