Adium now supports, and correctly responds to, URLs of the following type:
	aim:goim?screenname=xxx

where xxx is the name of an AIM user. In addition, support has been added, but not enabled,
for URLs of the following type:

	ymsgr:sendim?xxx

where xxx is the name of a Yahoo! Messenger user. I'm waiting for Yahoo to actually work to
enable the latter class of URLs.

There is also a minor change to the -preferedAccountForSendingContentOfType:toListObject:
method, whereupon an object of AIListObject is now treated properly, and not sent -accountID.