A second patch from retsam that I just finally got around to putting up. In his own words:
i fixed a bug that caused "Removed Idle" to show up in the File menu
when it wasn't supposed to and now auto away preferences dim when no away
messages are saved, and "No saved aways" appears in the popup. also, i
had something inside of a while loop that didn't need to be inside of it.

There you have it folks. :-)