Message History Display!
When a new conversation is opened, the last few lines of the previous conversation are displayed in the message window. Currently, you can only turn this on and off. More customization and prettiness to come later.

The standard message view displays these history lines as greyed out. The webkit message view doesn't do anything special with them.

Many more changes to come, later...