Write the icns file out on every launch, not just when the preference changes; this lets Adium's finder icon stay correct between version upgrades.
Note that we only need to do this when notification != nil OR when Adium has been upgraded since the last launch.  As I mentioned in the #warning I put next to it, it'd make sense to have Adium keep an internal note of its version and be able to send out a versionDidChange notification when it is launched with a new version (before upgrading the preference to the new version number).