Twelve hours later, revamped webkit plugin with preferences. :)  I'll post on the blog the details relevent to webkit authors sometime tomorrow.
Preferences are INCOMPLETE but include working preview (using a plist which is both localizable and can be overridden by MessageStyles to provide a custom preview conversation) and message style selection.
We now support variants.. they should show up in a submenu.  Haven't actually tried that, but I think the code's right ;)
MessageStyles are bundles which include an info.plist file.  The file's important, as it supplies the (localizable) display name of the style as well as potentially other information.  The only other information we draw from it at present is that styles supporting variants can specify a name for the main.css file (i.e. the no-variant version).
The file formally known as test.css is now main.css. Variants go in a Variants folder in Contents/Resources (everything else also goes in Contents/Resources).
A lot of the important functionality has been moved to the plugin to let code be shared properly between the preferences and the message view.

More stuff will be done with info.plist in the future, I imagine.. not quite sure what yet though :)