Added a warning dialog when a new custom plugin is detected in ~/Library/Application Support/Adium 2.0/Plugins.
The dialog tells the user they will crash and everything will expload,
and offers the option to continue or disable the plugin.  Disabling
moves it to a Plugins (Disabled) folder.  The dialog will only be
displayed once for each plugin.

This should help avoid the '0.53 is crashing!?' flood resulting from
re-enabling that folder with the old plugins in it (which haven't been
getting loaded up until now)... or atleast partially suppress it.

:)