Revamped adding a contact to a meta-contact. Has a text field which will filter results + the ability to add all filtered to a meta-contact.
Text field uses like, so wildcards work. (% and ?)

Switched setting the preferred meta-contact to meta.jsp.  Cleared up some room
for users, and moved the note information to the left.  Works nicely.

Insert some information keys by default; made editMeta and editUser behave better if there was no data in
information_keys. :)

Added better handling in saveNote in case of an error.
Switch index.jsp to use a left join instead of a subselect.  It seems like
this would be much faster, but in my testing it was about the same.