I'm fairly sure this fixes: - Jabber presence problems which were requiring a really ugly hack to work around - Improper ICQ authorization requests - Possibly AIM group deletion.
It's amazing how a subtle bug can cause a huge range of problems.  We were (and have been since the initial gaim threading attempts) setting our away status in the main thread.  It turns out that doing so causes all sorts of server commands to bounce around... and we were occassionally doing it the midst of the signon activity (in the Gaim thread).
This causes, in the case of Jabber (which was the easiest to work with, and tipped me off to the source of the problem), us to get Presence notifications from our buddies immediately -before- we receive our permit/deny list.... which leads to us ignoring all their online statuses.
The workaround before was to manually request the roster as soon as we thought we were connected... Gaim didn't have to do this, though, so something had to be wrong on our side.  A whole lot of toying with things made me realize that it only happened when I was away... and then it became clear from NSRunLoop logging that the serv_set_away was being called from the wrong thread, triggering callbacks and such in the  main thread concurrent with Gaim activity.
The Jabber bug is fixed.  We never had a reliable way to reproduce the ICQ and AIM signon bugs, but I strongly suspect they were caused by this same issue (which was occurring in base Gaim Plugin code, in CBGaimAccount).
The resulting code is cleaner in terms of shuffling Gaim responsibilities into SLGaimCocoaAdapter, as well :)

While working on this, I also redid our gaim-requested timed callbacks to use CFTimers inserted into the same run loop mode as the immediate callbacks, which should help ensure they are called in the right order, as well.