Icons are now used to display status in tabs, replacing the flawed text coloring we used previously.   May take a little bit to get used to, but I think everyone will like this in the end :).  It makes seeing status a whole lot easier, and definitely gives us a more polished look.
The icons are a lot easier to see, and will be easier to use since they aren't based entirely on
colors.  For now the images are hard coded, but we'll make them customizable packs later on.  Not happy
with my new content icon, but the rest are decent enough at this point.

Still need to clean some stuff up and hook up the stranger status state.