Okay, I give up.  I've wasted way too much time trying to move the webkit code from the plugin to the controller in order to fix these two issues and make the code a lot cleaner looking.  I'm noting the two issues I was trying to fix incase someone else wants to tackle the impossible.
All it would take is moving the code that effects the webkit view into
the controller.  Then, the preference pane could be switched to use a
message view controller instead of directly communicating with with
webkit view.

There's a third issue that involves not using the webkit view to save
preferences, since that doesn't work with Adium's multiple user
system... but we're dropping that when we drop 10.2 so I could
care less about that one :)