Changes to Adium to use the new libgaim signal for setting status on connect.  This should (I hope) fix the issues with contact list corruption and deletion when connecting.
Previously, Adium was setting status values (away, idle, profile, buddy icon) too early in the connection process.  On connect, the
server sends libgaim the server-side buddy list.  Libgaim stores the server list locally, and then when any changes to the contact list
are made, it runs through the local list and server list, comparing the two and syncing anything that is different.

What was causing us problems is that some of the server-side status (especially buddy icons) are stored using the same mechanism as the
contact list.  Adium would change the icon while the account was connecting, which would attempt to store the icon server-side.  In
storing the icon, libgaim would perform a list sync (since icons are stored with the same mechanism as the contact list).  This list sync,
if performed while libgaim was in the middle of downloading the server-list (or slightly before), would cause libgaim to start syncing a
partial or empty contact list with the server, and deletion of contacts.

It's possible that setting these status values early is also responsible for the issues with ICQ authorization, but I cannot confirm
that.  I have not tested and confirmed that our MSN display name is still working with these changes, but it should be easy to fix.

I'm not 100% familiar with how libgaim works, so forgive me if any of the above is incorrect.  But I'm feeling confident that this was the
cause of trouble (Cross your fingers and hope with me! :) ).  Unfortunatly, since this is so hard to reproduce, I can't be 100% sure that
the issue is resolved.  Best chance to reproduce it is connecting while away and idle, after having just changed your user icon and
profile.