Fixed an error in details.jsp which caused problems with a box.
Made index.jsp use a URLEncoder method instead of doing it manually and
poorly.
Fix the calendar icon to point to the right place.

Get rid of parser.pl, because with the Adium 1.x log importers, it's
unnecessary.  Hasn't been updated in a while, anyway.

Renamed parser-2.pl to parser.pl.

No longer redirect to the meta page instead of the users page after updating
the login users.