AIAccount changes for the ESObjectWithStatus code.  Forgot to mention in the previous commit:
When setting status on an object, notify: may now be NotifyNever, NotifyNow, or NotifyLater.  NotifyNow is the same as YES from before (and passing YES remains valid).  NotifyLater is the same as NO from before (and passing NO remains valid).  Passing NotifyNever should be done when the status object is being used only for internal bookkeeping, for efficiency.