Patch from Steven Tamm (steventamm) with some modifications by me to add a Keys preference for the behavior of the Home/End keys.
The default of home/end scrolling the message view is retained; with "[ ] 'Home' and 'End' apply to the text input area" checked, home/end and shift+home/shift+end apply to the text in the normal OS X behavior.

I'm open to making this the default behavior and rewording the preference to something along the lines of "[ ] 'Home and 'End' scroll the message view".  Let me know what you think.