We now have a manual ordering system which doesn't require loading ordering information from a monolithic contact list.plist.
A ContainingObject now knows the max and min ordering indexes inside of it; the contact controller uses this information along with the ordering index as before to make new ordering determinations when the user changes the order.  ListObjects are responsible for saving and loading their own ordering information.
The contactController keeps track of what the next unique ordering index will be, as this algorithm assumes that no two contacts have the same ordering index (since (X-X)/2 is 0).