Adding three color themes and one layout which I like; I prose including these by default along with others.  I'm not married to the Aqua Grid, but it seemed simple and clean; I think the 2 and 3 tone status coloring ones look good. Group Bubbles - Centered.ListLayout is my current list layout and I think it's a good example of how clean Mockie (erm, Group Bubbles) can look.
Perhaps we should have a forum thread of basic (that is, ones that look good and aren't dependent at all on having your desktop be a certain way for them to be nice) themes and layouts for potential inclusion in the default build?  At the least, I'd love it if any other developers who like their list set up would email them to either me or Adam.  I think we should have one or two for each of the window styles, and 5 or 6 coloring themes.