Fixed an issue with certain fonts/sizes not causing an update of the font height cache (resulting in extended status drawing at an incorrect vertical position).  Apparently font objects are being reused, so an equality check in out setFont: method was preventing the cache update code from running (but, since the original font object had been changed elsewhere, the font seemed to have been updated correctly on the next display).
Also added a plugin to move extended status from the status arrays into the display arrays.  This gives us a bit more control over display,
allowing the status to be prefixed with idle time, stripped of line breaks and cropped to a reasonable length (sizing giant paragraphs is slow
and pointless when we can only see the first 50 characters or so).