Added an option for Applescript filters, specifiable on a per-script basis via the info.plist, <key>Requires User Interaction</key>
The default is NO.  If this key is specified to be YES, <true/>, then the script is run in the main thread instead of the filter thread.

When NO, attempts to intereact with the Adium UI are almost guaranteed to crash. (Messages sent via Adium's own applescript commands are fine; we're talking UI scripting and display dialogue and such here).
When YES, such scripts work as expected.  Note that setting YES means that Adium will wait to finish the script before returning to normally scheduled programming, so it shouldn't be set unless there's a good reason to do so.