- (NSImage *)imageByScalingToSize:(NSSize)size fraction:(float)delta flipImage:(BOOL)flipImage is now - (NSImage *)imageByScalingToSize:(NSSize)size fraction:(float)delta flipImage:(BOOL)flipImage proportionally:(BOOL)proportionally
if proportionally is NO, the method behaves as before, drawing the image at size 'size' regardless of previous proportions.
if proportionally is YES, size will be used as a maximum size; if height != width, the smaller of the two will be scaled to within size by the same proportion that the larger one changes.