Added setAllowFilterThreadAutoreleasePoolRefresh:(BOOL)flag to the content controller, and made use of it in the applescript filters plugin.
The problem is that executing an applescript doesn't leave the invocation but -does- span multiple run loops... so timers have a chance to fire in the middle of the current invocation.
That's fine in some circumstances, but when a timer fires, release the current autorelease pool, and then control returns to the middle of a method which is expecting that its autoreleased variables are still around, well, that's a mess.

I played with several solutions to this problem, but it's difficult to find a clean one.  You can't use nested autorelease pools to fix this, because when the outer one is released, the inner one is automatically released as well.

The applescript filters plugin now tells the contentController not to allow the autorelease pools to be released and recreated, sending NO before execution and YES afterwards to reenable it.  I don't believe this measure should be needed anywhere else at present, but watch for odd autorelease crashes just in case (latex plugin, for example, might see the same problems).