Some OCD safety based on two very rare crashes reports submitted recently.
In AIOutlineView, it appears that setFrame: can lead to tile which can lead to setFrame: which can lead to tile which can..... recursive call protection added.
In AIMessageEntryTextView, the docs say that setAllowsUndo: is supported by all versions of OS X.. but someone crashes on a does not respond to selector exception.  checking against respondsToSelector: can't hurt.