Thread lock this, Applescript.
We now have two sets of filters, contentFilters and threadedContentFilters.  threadedContentFilters will only be applied when the calling method uses the filterAttributedString: call which passes a target and selector.
This means that all the other filters now run in the main thread; while this means that signing on 9 accounts at once is a little jerkier for the first moment as tons of strings are filtered, it provides about the same performance with none of the nasty threading interactions which were necessary to be able to give a tooltip (or other item which needs an immediate return) a filtered attributedString.