Major dock icon preferences updates and simplification (using the new imageGridView class).  Clean code is happy code!
I've changed the dock preferences to animate when the icons are hovered, but it'd be easy to change back if everyone likes it better the old way (or with no
animation at all if we want).  Because the image grid view handles all the cursor tracking we can do anything with minor changes delegate side in the
preference code.