Fixed issues with generic idle times, specifically with regards to MSN's idle state.  MSN idle is now set and unset properly for contacts.
So far as I can tell, MSN does not support "idle" as a state separate from an away state... so an idle contact who is also away only shows up as away.  The official client, at least for Mac, behaves the same way -- in fact, the official client just says "away" when a contact is idle... go figure.