The Finder-style name searching in AIOutlineView is now implemented via the delegate method  (void)outlineView:(NSOutlineView *)outlineView userDidTypeString:(NSString *)inputString matchTargetNumber:(int)ta rgetNumber;
The change is the addition of the matchTargetNumber: parameter.  The user can press tab to increase this number -- so the first call as the user types will have a targetNumber of 1, and if as the user presses tab the targetNumber increases.  The delegate can use this information if desired to search for the second hit, the third hit, and so on.