Another type select patch from Ken Ferry, who writes:
It isn't always appropriate to search every column of a table.  For example, in Preferences > Accounts, the images corresponding to different kinds of accounts have descriptions starting with "NSImage", so 'N' matches every row.

The patch specifies search columns in 3 spots where I found undesirable behavior.  There may be others, but it's sometimes difficult for me to find tables in the interface based on the source.  Anyway, it's always an easy fix.

Thanks, Ken :)