Corrected a few problems with the upgrade to PostgreSQL 8.0, which shouldn't affect any other users at all.
These problems include the chats page not loading correctly and the stats page
not loading correctly.

Fixed a couple problems with the log parsing script importing too much data.

Saving a form was broken because of a stupid mistake.  Now it's not.

Updated Axamol to r1035.
Switched ConnectionManager.java to Axamol's handling of it.

Rearranged the website layout without affecting content.