If multiple packs are enabled, and multiple packs have matches for an emoticon [e.g. :) is in both the AIM and the Yahoo! packs], then the service of the destination contact, or the contact for which filtering is occuring, is consulted.  The :) from AIM would be used for an AIM contact while the one from the Yahoo! pack would be used for a Yahoo contact.
Ordering is still respected, so multiple Yahoo packs for example can be enabled and will be consulted in order.  Packs with no service association are preferred to those on the wrong service, so the "Crazy Emoticons" :) would be used over the AIM :) for a Yahoo contact, whereas the ordering of the two in the emoticon customization would be respected for an AIM contact.

I believe this answers the relatively common feature request for emoticons to appropriate for the service while still providing the same functionality for people who use multiple other packs, order packs to prioritize them, and so forth.