Moved the -[AIBookmarksImporterController buildBookmarksMenuIfNecessaryThread] to validateMenuItem:
menuNeedsUpdate is called with nearly every keystroke for an item in the main menu.... it's a bad idea to be detaching a new thread and checking the bookmarks every time a key is pressed.  Since validateMenuItem is only called when the item is going to be displayed, this is more appropriate.

We're detaching a thread, so either way the update if one is made isn't going to be visible until the _next_ time the menu is accessed.