When the entire contents of a text view change and it is sending text, immediately proces the typing so that the not-typing message is sent if appropriate.
When the entire contents change and it is not currently sending text -- which means that the user did a [select-all,delete], [cut], or [paste] operation, in general -- delay 2 seconds before processing that typing notification.  This means Adium now has the same typing notification behavior as iChat and, by extension, plays much better with clients such as iChat which make a huge annoying deal out of the typing notification... for example, if you select all, delete, and then start typing within 2 seconds, the typing 'bubble' in iChat simply remains rather than flickering out and then back in.